full comment system updated
updated comment system

+++ tamim-web2.1/tamim-web2/database/migrations/2016_08_18_152344_create_laravellikecomment_total_likes_table.php
@@ -13,10 +13,10 @@ class CreateLaravellikecommentTotalLikesTable extends Migration
     public function up()
     {
         Schema::create('laravellikecomment_total_likes', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('item_id'); // ModelName_modelId
-            $table->integer('total_like');
-            $table->integer('total_dislike');
+            $table->bigInteger('total_like');
+            $table->bigInteger('total_dislike');
             $table->timestamps();
         });
     }
