+++ tamim-web2.1/tamim-web2/database/migrations/2016_08_18_152401_create_laravellikecomment_likes_table.php
@@ -13,11 +13,12 @@ class CreateLaravellikecommentLikesTable extends Migration
     public function up()
     {
         Schema::create('laravellikecomment_likes', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('user_id');
             $table->string('item_id'); // ModelName_modelId
-            $table->smallInteger('vote');
+            $table->bigInteger('vote');
             $table->timestamps();
+
         });
     }
